Fixed the overwrite problem when devices offline. Refactor: Rename `showlast` package to `manageitem` and update quantity type to Long.
- Renamed the package `eu.jelinek.hranolky.ui.showlast` to `eu.jelinek.hranolky.ui.manageitem`.
- Updated all relevant files and imports to reflect this change.
- Changed the `quantity` field type from `Int` to `Long` in `WarehouseSlot`, `FirestoreSlot`, `SlotSum`, and related use cases and repository methods. This allows for larger quantities.
- Improved the logic for creating new slots in `SlotRepositoryImpl` to use Firestore transactions and prevent overwriting existing slots.
- Updated the header text in `ManageItemScreen` to dynamically display "Spárovka" or "Hranolek" based on the `slotId`.
- Incremented app version code to 3 and version name to "1.2".

+++ app/src/main/java/eu/jelinek/hranolky/data/SlotRepository.kt
@@ -2,14 +2,14 @@ package eu.jelinek.hranolky.data
 
 import eu.jelinek.hranolky.model.SlotAction
 import eu.jelinek.hranolky.model.WarehouseSlot
-import eu.jelinek.hranolky.ui.showlast.ActionType
+import eu.jelinek.hranolky.ui.manageitem.ActionType
 import kotlinx.coroutines.flow.Flow
 
 interface SlotRepository {
     fun getSlot(slotId: String): Flow<WarehouseSlot?>
     fun getSlotActions(slotId: String): Flow<List<SlotAction>>
-    suspend fun addSlotAction(slotId: String, actionType: ActionType, quantity: Long, currentQuantity: Int, deviceId: String)
-    suspend fun createNewSlot(slotId: String, quantity: Int)
+    suspend fun addSlotAction(slotId: String, actionType: ActionType, quantity: Long, currentQuantity: Long, deviceId: String)
+    suspend fun createNewSlot(slotId: String, quantity: Long)
     fun getLastModifiedSlots(): Flow<LastModifiedSlots>
     fun getAllSlots(): Flow<List<WarehouseSlot>>
 }
